Use spawnSync so we can support older IAR versions
Older IAR versions do not support the --version argument. So they will
return an errorcode and thus execSync fails to execute. spawnSync does not
throw an exception, the error code is stored in the return object. This
way we can still parse the version number as it is also printed when an
invalid argument is given to the compiler.

Signed-off-by: Philip Luyckx <philip.luyckx@gmail.com>