Fix issue this pointer in handler call
We assign the this pointer to the function object. When adding multiple,
different handlers, they overwrite the this pointer. Use the handler object
that is introduced in a couple of commits back to solve this. We create
for each handler its own object and assign the this pointer to that object.