Split command implementation in base and input command
The commands we have implemented are always commands that trigger a user
input. However, we will need also commands that do not need a user input.
Therefore we have to split the current implementation into a `base` class,
which implements all methods that are common for a Command. The new class
`CommandWithInput` is a command that needs user input.
Other type of commands can reuse the `CommandBase` class.