Add defines to settings
Depending on the compiler there are compiler specific intrinsics. This is
hard to maintain in the plugin or keep the plugin up-to-date. To solve this
the new setting can be used. The user can adapt the define list and add
compiler intrinsics to the define list so VSCode does not add unwanted
items to the `Problem` list.