Correct from path to calculate relative
the uri.path returns a path in the form `/c:/path/to/`. Becasuse of the
first `/`, the relative function will add an extra `c:\`, so you get
`c:\c:\path\to\`. Now a lot of `..\` are added to the relative path. Using
`fsPath` solves this and a true system path is returned: `c:\path\to`.