#51: Add central `CommandManager`
We need a central manager to find specific commands. We cannot use the
`executeCommand` method from `vscode` because this is an async function and
is executed later. Because of this we cannot use this to fetch settings
for the process executable. So we have to find and call the command by
ourself. This module will make this possible.