Let SingletonModel value be set via promise
Every use of SingletonModel creates a promise that sets the value of the
model when resolved. This commit lets such promises be stored and
managed by the SingletonModel to:
 * Ensure values from previous promises are discarded after creating a
new one
 * Let others find such promises and await them

For example, if the extension is initialized as a result of the user
calling 'createProject', that command needs to be able to wait for
an ExtendedWorkbench to be loaded before it can run.